requirements.txt
chromadb>=0.3.29
langchain>=0.2.0,<0.3.0
langchain-community>=0.2.0,<0.3.0
langchain-openai>=0.1.0,<0.2.0
langchain-chroma>=0.1.0,<0.2.0
langchain-core>=0.2.40,<0.3.0
langsmith>=0.1.112,<0.2.0
openai>=1.12.0
tiktoken>=0.5.2
unstructured>=0.8.8
httpx>=0.26.0
nltk>=3.8.1
git+https://github.com/jufist/oliver-framework.git
python-dotenv>=1.0.0
PyYAML>=6.0
